refactor: 테스트 파일 폴더에 넣기
